Add a new wrapper for calling tardis
The wrapper points the tardis logger to a file inside
TardisWrapper._log_dir
The wrapper is expected to be called with a callback function as
argument. This function is called with a ConfigNamespace as argument and
is expected to return a ConfigNamespace. This allows complete freedom
when manipulating the config in subsequent runs, keeping the wrapper
free from any implementation.

TardisWrapper.run_tardis(config) is expected to be modified if more
complicated evaluation of Tardis is desired (See TInnerWrapper).