Add basic components to construct chains
Two building blocks for constructing a chain are introduced.
Each operation is represented by a Link object. These objects can be
chained with the Chain object. For examples look into test_base.py
A simple Link class looks like this:
```
class Add(Link):
    inputs = ('x',)
    outputs = ('y',)

    def __init__(self, a):
        self._a = a

    def calculate(self, x):
        return x + self._a
```
Several instances of Link can be chained with Chain where the arguments
to Chain's constructor get evaluated from left to right. The keyword
argument 'breakable' allows the raising of BreakChainException to
terminate execution of the current Chain which sets all unset outputs to
None.

A Chain propagetes it's state with a dictionary where new outputs are
added as new keys. This should be replaced by a custom Container but for
now it works.