Add tools for statistical analysis
The prior is designed to be calculated before the Tardis run. CheckPrior
is used to skip the tardis evaluation in case the prior is INVALID.
This is done by raising BreakChainException which breaks the innermost
Chain. Thus CheckPrior should always be called before Tardis and at
least on the same level or higher.

The only loglikelihood implemented so far is SSum. More to come here.
The posterior just adds loglikelihood and logprior. If loglikelihood is
None(because the tardis-run was skipped), only the prior is returned,
which should be INVALID in that case.