Add a simple composition of a typical tardis run
This is just a shortcut which abstracts a typical tardis evaluation with
the SSum as likelihood.

SimpleTardis can be used as an example how to define basic compositions,
for example to use only one 'component', when defining a custom chain,
for the operation of calculating the flux from the model output.