Add ParameterContainer and Parameter class
ParameterContainer is a simple Container to be initialized with a set of
Parameters. Parameters that are calculated from other Parameters are
supported. CalculatedParameter is supposed to handle that but this is
not yet implemented. OverFlowParameter works though.

Setting parameter values can be done in different ways.
There is Parameter.value = 123 to set a single value or
ParameterContainer.<Parameter Name>.value = 123 is the same.

For setting multiple variables, beginning with the first defined
Parameter and skipping CalculatedParameters there is
ParameterContainer.values = [1,2,3,4]