Rework Parameter structure
The default Parameter is static and it's value can't be changed.
The DynamicParameter replaces the old Parameter and provides the
functionality to map an input of [0,1] onto the output [bounds[0],
bounds[1]]. The mapping function can be changed with the keyword
argument 'transformation'.

A DependentParameter class was introduced to handle Parameters depending
on other Parameters. These will be called last.
Two Parameters depending on each other aren't supported.