Remove Exception raising for illegal parameters
Raising Exceptions when setting parameters was a bad idea. This behavior
changed to return np.nan instead of raising an exception.
This allows for debugging (which parameter got an illegal value?) and
simplifies working with parameters in dalek.

Because parameters can now be nan an additional check is added to the
Prior to cancel the run in that case.

There might have been a bug with OverFlowParameters where they would
include their previous value during calculation of the new value.
This should be fixed now and a simple test for that case was added.